refactor: add centered layout for login page
