+++ src/app/App.tsx
@@ -3,13 +3,16 @@ import { BrowserRouter } from 'react-router-dom';
 import AppRoutes from './AppRoutes';
 import { ThemeProvider } from '@mui/material/styles';
 import { appTheme } from '@shared/themes/theme';
+import { CssBaseline } from '@mui/material';
 
 const App: FC = () => {
   return (
     <BrowserRouter>
-      <ThemeProvider theme={appTheme}>
-        <AppRoutes />
-      </ThemeProvider>
+      <CssBaseline>
+        <ThemeProvider theme={appTheme}>
+          <AppRoutes />
+        </ThemeProvider>
+      </CssBaseline>
     </BrowserRouter>
   );
 };
