--- src/pages/LoginPage/LoginPage.module.css
@@ -1,97 +1,0 @@
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
-    background: #f0f0f0;
-    font-family: 'Arial', sans-serif;
-}
-
-.form {
-    background: white;
-    border-radius: 8px;
-    padding: 40px;
-    box-shadow: 0 0 20px rgba(0, 0, 0, 0.1);
-    width: 100%;
-    max-width: 400px;
-    text-align: center;
-}
-
-.title {
-    font-size: 2.5rem;
-    color: #333;
-    margin-bottom: 20px;
-}
-
-.image {
-    width: 80px;
-    border-radius: 50%;
-}
-
-.input {
-    width: calc(100% - 24px);
-    padding: 12px;
-    margin: 15px 0;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    font-size: 1rem;
-    color: #555;
-    box-sizing: border-box;
-}
-
-.checkbox {
-    display: flex;
-    justify-content: left;
-    align-items: center;
-    margin: 10px 0 30px;
-    padding-left: 10px;
-}
-
-.checkbox input {
-    margin-right: 10px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.button {
-    width: 100%;
-    padding: 12px;
-    background: #1976D2;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    font-size: 1rem;
-    cursor: pointer;
-    transition: background-color 0.6s;
-}
-
-.button:hover {
-    background-color: #285596;
-}
-
-/* Стили для успешного сообщения */
-.successMessage {
-    background-color: #2ecc71;
-    color: white;
-    padding: 15px;
-    border-radius: 5px;
-    margin-top: 15px;
-}
-
-/* Стили для сообщения об ошибке */
-.errorMessage {
-    background-color: #e74c3c;
-    color: white;
-    padding: 15px;
-    border-radius: 5px;
-    margin-top: 15px;
-}
-
-@media only screen and (max-width: 600px) {
-    .form {
-        width: 60%;
-        max-width: 100%;
-    }
-    .image {
-        width: 60px;
-    }
-}
